// client/src/repo/lottie-graphql/__generated__/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** The javascript `Date` as string. Type represents date and time as the ISO Date string. */
  DateTime: { input: any; output: any; }
  /** A field containing either a integer or a string value */
  IntOrString: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
  /** Point scalar type */
  Point: { input: any; output: any; }
  /** Query data sort options scalar type */
  QuerySortOptions: { input: any; output: any; }
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  /** When paginating forwards, the cursor to continue. */
  endCursor?: Maybe<Scalars['String']['output']>;
  /** When paginating forwards, are there more items?. */
  hasNextPage: Scalars['Boolean']['output'];
  /** When paginating backwards, are there more items?. */
  hasPreviousPage: Scalars['Boolean']['output'];
  /** When paginating backwards, the cursor to continue. */
  startCursor?: Maybe<Scalars['String']['output']>;
};
// ...
export type PublicAnimation = {
  __typename?: 'PublicAnimation';
  bgColor?: Maybe<Scalars['String']['output']>;
  comments?: Maybe<Array<PublicComment>>;
  commentsCount: Scalars['Int']['output'];
  createdAt: Scalars['DateTime']['output'];
  createdBy?: Maybe<User>;
  createdByUserId: Scalars['ID']['output'];
  description?: Maybe<Scalars['String']['output']>;
  downloads?: Maybe<Scalars['Float']['output']>;
  gifFileSize?: Maybe<Scalars['String']['output']>;
  gifUrl?: Maybe<Scalars['String']['output']>;
  id: Scalars['Int']['output'];
  imageFileSize?: Maybe<Scalars['Int']['output']>;
  imageFrame?: Maybe<Scalars['String']['output']>;
  imageUrl?: Maybe<Scalars['String']['output']>;
  isCanvaCompatible?: Maybe<Scalars['Boolean']['output']>;
  isLiked: Scalars['Boolean']['output'];
  jsonUrl?: Maybe<Scalars['String']['output']>;
  likesCount: Scalars['Int']['output'];
  lottieFileSize?: Maybe<Scalars['Int']['output']>;
  lottieFileType?: Maybe<Scalars['String']['output']>;
  lottieUrl?: Maybe<Scalars['String']['output']>;
  lottieVersion?: Maybe<Scalars['String']['output']>;
  name: Scalars['String']['output'];
  publishedAt?: Maybe<Scalars['DateTime']['output']>;
  slug: Scalars['String']['output'];
  sourceFileName?: Maybe<Scalars['String']['output']>;
  sourceFileSize?: Maybe<Scalars['Int']['output']>;
  sourceFileType?: Maybe<Scalars['String']['output']>;
  sourceFileUrl?: Maybe<Scalars['String']['output']>;
  sourceName?: Maybe<Scalars['String']['output']>;
  sourceVersion?: Maybe<Scalars['String']['output']>;
  speed?: Maybe<Scalars['Float']['output']>;
  status?: Maybe<Scalars['Int']['output']>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  url?: Maybe<Scalars['String']['output']>;
  videoFileSize?: Maybe<Scalars['Int']['output']>;
  videoUrl?: Maybe<Scalars['String']['output']>;
};

export type PublicAnimationConnection = {
  __typename?: 'PublicAnimationConnection';
  /** A list edges. */
  edges: Array<PublicAnimationEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars['Int']['output'];
};

export type PublicAnimationDeleteResponse = {
  __typename?: 'PublicAnimationDeleteResponse';
  message: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type PublicAnimationEdge = {
  __typename?: 'PublicAnimationEdge';
  /** A cursor for use in pagination. */
  cursor: Scalars['String']['output'];
  /** The item at the end of the edge. */
  node: PublicAnimation;
};
// ...
export type PublicComment = {
  __typename?: 'PublicComment';
  content: Scalars['String']['output'];
  createdAt: Scalars['DateTime']['output'];
  frame?: Maybe<Scalars['Float']['output']>;
  id: Scalars['Int']['output'];
  isResolved: Scalars['Boolean']['output'];
  marker?: Maybe<Scalars['Point']['output']>;
  parentId?: Maybe<Scalars['Int']['output']>;
  replies?: Maybe<Array<PublicComment>>;
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  user?: Maybe<User>;
  userId: Scalars['ID']['output'];
};
// ...
export type Query = {
// ...
  featuredPublicAnimations: PublicAnimationConnection;
// ...
};
// ...
export type User = {
  __typename?: 'User';
  achievements?: Maybe<UserAchievementConnection>;
  /** The date/time when the user authenticated the current session. */
  authenticatedAt: Scalars['DateTime']['output'];
  /** The user's profile picture. */
  avatarUrl: Scalars['String']['output'];
  /** The user's Behance username. */
  behanceUsername?: Maybe<Scalars['String']['output']>;
  /** The user's short Bio. */
  bio?: Maybe<Scalars['String']['output']>;
  /** The user's City of residence. */
  city?: Maybe<Scalars['String']['output']>;
  /** The user's Country of residence. */
  country?: Maybe<Scalars['String']['output']>;
  /** The date/time of account creation. */
  createdAt?: Maybe<Scalars['DateTime']['output']>;
  /** The user's Dribbble username. */
  dribbbleUsername?: Maybe<Scalars['String']['output']>;
  /** The user's email address. */
  email?: Maybe<Scalars['String']['output']>;
  /** True if the user's current email address has been verified. */
  emailVerified?: Maybe<Scalars['Boolean']['output']>;
  /** True if the user has agreed to receive marketing emails. */
  enableMarketingEmails?: Maybe<Scalars['Boolean']['output']>;
  /** The user's first name. */
  firstName: Scalars['String']['output'];
  /** The user's Github username. */
  githubUsername?: Maybe<Scalars['String']['output']>;
  hasSlackNotificationEnabled?: Maybe<Scalars['Boolean']['output']>;
  /** The user's LottieFiles account's unique ID. */
  id: Scalars['String']['output'];
  /** The user's Instagram username. */
  instagramUsername?: Maybe<Scalars['String']['output']>;
  /** True if the user is available for work. */
  isHireable: Scalars['Boolean']['output'];
  /** @deprecated Legacy field, will always return false. */
  isPro: Scalars['Boolean']['output'];
  /** The city where the user last logged in from. */
  lastLoggedInFromCity?: Maybe<Scalars['String']['output']>;
  /** The country where the user last logged in from. */
  lastLoggedInFromCountry?: Maybe<Scalars['String']['output']>;
  /** The timezone where the user last logged in from. */
  lastLoggedInTimezone?: Maybe<Scalars['String']['output']>;
  /** The user's last name. */
  lastName?: Maybe<Scalars['String']['output']>;
  /** For use in the internal migration process. It will contain the ID of the user in the legacy WEB DB if the user is migrated. Otherwise, it will return an INT hashed from the user's Kratos ID. */
  legacyWebUserId: Scalars['Int']['output'];
  /** The user's Linkedin username. */
  linkedinUsername?: Maybe<Scalars['String']['output']>;
  /** The user's locale code. */
  locale?: Maybe<Scalars['String']['output']>;
  /** @deprecated Use `city` instead. */
  location?: Maybe<Scalars['String']['output']>;
  /** @deprecated Use `firstName` instead. */
  name: Scalars['String']['output'];
  publicAnimations?: Maybe<PublicAnimationConnection>;
  stats?: Maybe<UserStats>;
  /** The user's Twitter username. */
  twitterUsername?: Maybe<Scalars['String']['output']>;
  /** If the user is ongoing an email change process, this is the new email that the user is changing to. Otherwise it will be `null`. */
  unconfirmedNewEmail?: Maybe<Scalars['String']['output']>;
  /** The date/time of last account update. */
  updatedAt?: Maybe<Scalars['DateTime']['output']>;
  /** @deprecated Use `username` instead. */
  url: Scalars['String']['output'];
  /** The segments this user belongs to. */
  userSegments?: Maybe<Array<UserSegment>>;
  /** The user's unique username. */
  username: Scalars['String']['output'];
  /** The user's personal website. */
  website?: Maybe<Scalars['String']['output']>;
};
// ...
export type FeaturedPublicAnimationsQueryVariables = Exact<{
  after?: InputMaybe<Scalars['String']['input']>;
}>;


export type FeaturedPublicAnimationsQuery = { __typename?: 'Query', featuredPublicAnimations: { __typename?: 'PublicAnimationConnection', edges: Array<{ __typename?: 'PublicAnimationEdge', node: { __typename?: 'PublicAnimation', id: number, slug: string, lottieUrl?: string | null, name: string } }>, pageInfo: { __typename?: 'PageInfo', hasNextPage: boolean, endCursor?: string | null } } };